requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
beanie==1.26.0
pymongo==4.8.0
python-dotenv==1.0.0
httpx-oauth==0.13.1
pyjwt==2.8.0
python-multipart==0.0.6